requirements.txt
requests
rapidocr-onnxruntime
pillow

